'Charities Accepting Dogecoin' Article
There's *heaps* of charities that accept Dogecoin, and vetting each one with the aim of providing a list would be a mammoth task. Here I've mentioned a few of the bigger charities in addition to The Giving Block and formed a more generalised response, but it could also be great to single-out some smaller charities that would benefit from a mention here. Please amend / add / edit / delete wherever. :)