Privacy Policy Fixes (Needs Improvements)
This privacy policy need work, guys. I am not a privacy policy expert, I did remove the "company is Dogecoin" line as it is inappropriate and might be misinterpreted/misused.

But the policy as a whole needs rewriting I suppose, even though this being a static website hosted on Github, it should end up being quite streamlined.