Contributing Guidelines - First Draft
Hello folks,

here is a suggestion for the contributing guidelines to this repository. This is obviously based on the guidelines we use for Core (https://github.com/dogecoin/dogecoin/blob/master/CONTRIBUTING.md), with changes to fit the specifics of a website repository.

The purpose of these guidelines is:

- Add transparency to the review process 
- Have a written document *future* actions can be matched against
- Introduce simple rules about PRs (mostly the fact PRs must be focused, whenever possible, and not make tracking history of the content difficult)
- Help to avoid that no individual, group of individuals, or organization can take full control over the repository (these guidelines can only do so much in that sense, since other factors like ownership of the domain come into play)
- Help to prevent blunders that we all make once in a while, but that can have serious repercussions for the entire community when they are published on Dogecoin.com

Please note, to avoid any tension: this is not just a response to specific incident, or meant to be an attack against any individual or group, but something that I think might help fulfill the mission of this website the way many of you have envisioned it even before it was pushed live on Dogecoin.com. I had the privilege to work with Richard, Tim and others on the baseline of the Dogepedia, and our purpose was to create something upon which other shibes would expand upon for years to come. And I know this sentiment was shared also by all contributors who have worked so incredibly hard on this website, from within and without the Foundation, each according to their possibilities. These guidelines are also to protect their efforts.

This is a first draft, to kickstart the discussion.